Refactor code and add correct version to build.gradle
Refactored the code to prevent possible compatibility issues with future mods using the same package, added correct version to Build.Gradle
