+++ build.gradle
@@ -11,8 +11,8 @@ apply plugin: 'net.minecraftforge.gradle.forge'
 //Only edit below this line, the above code adds and enables the necessary things for Forge to be setup.
 
 
-version = "1.0"
-group = "com.example" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
+version = "1.6"
+group = "io.github.vladocc" // http://maven.apache.org/guides/mini/guide-naming-conventions.html
 archivesBaseName = "BackToHome"
 
 repositories {
@@ -81,7 +81,3 @@ processResources {
         exclude 'mcmod.info'
     }
 }
-
-
-
-
