+++ src/main/java/io/github/vladocc/getbacktohome/ModInfo.java
@@ -1,4 +1,4 @@
-package com.example.examplemod;
+package io.github.vladocc.getbacktohome;
 
 /**
  * Created by Voyager on 07.05.2018.
