+++ src/main/java/io/github/vladocc/getbacktohome/commands/DeleteHomeCommand.java
@@ -1,42 +1,34 @@
-package com.example.examplemod.commands;
+package io.github.vladocc.getbacktohome.commands;
 
-import com.example.examplemod.HomeController;
-import com.example.examplemod.HomeInfo;
-import com.example.examplemod.HomeProvider;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
-import net.minecraft.entity.Entity;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.text.TextComponentString;
-import net.minecraft.world.storage.WorldInfo;
-import net.minecraftforge.common.DimensionManager;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Created by Voyager on 06.05.2018.
+ * Created by Voyager on 07.05.2018.
  */
-public class SetHomeCommand implements ICommand {
+public class DeleteHomeCommand implements ICommand {
 
     ArrayList<String> allias = new ArrayList<>();
 
-    public SetHomeCommand() {
-        allias.add("sethome");
+    public DeleteHomeCommand() {
+        allias.add("removehome");
     }
 
     @Override
     public String getName() {
-        return "createhome";
+        return "deletehome";
     }
 
     @Override
     public String getUsage(ICommandSender sender) {
-        return "createhome";
+        return "deletehome";
     }
 
     @Override
@@ -46,8 +38,7 @@ public class SetHomeCommand implements ICommand {
 
     @Override
     public void execute(MinecraftServer server, ICommandSender sender, String[] args) throws CommandException {
-        Entity entity = sender.getCommandSenderEntity();
-        HomeController.setHome(entity);
+
     }
 
     @Override
