+++ src/main/java/io/github/vladocc/getbacktohome/commands/MoveBackCommand.java
@@ -1,34 +1,38 @@
-package com.example.examplemod.commands;
+package io.github.vladocc.getbacktohome.commands;
 
+import io.github.vladocc.getbacktohome.HomeController;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
+import net.minecraft.entity.Entity;
+import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.text.TextComponentString;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
 import java.util.List;
 
 /**
- * Created by Voyager on 07.05.2018.
+ * Created by Voyager on 06.05.2018.
  */
-public class DeleteHomeCommand implements ICommand {
+public class MoveBackCommand implements ICommand {
 
     ArrayList<String> allias = new ArrayList<>();
 
-    public DeleteHomeCommand() {
-        allias.add("removehome");
+    public MoveBackCommand() {
+        allias.add("back");
     }
 
     @Override
     public String getName() {
-        return "deletehome";
+        return "home";
     }
 
     @Override
     public String getUsage(ICommandSender sender) {
-        return "deletehome";
+        return "home";
     }
 
     @Override
@@ -38,7 +42,8 @@ public class DeleteHomeCommand implements ICommand {
 
     @Override
     public void execute(MinecraftServer server, ICommandSender sender, String[] args) throws CommandException {
-
+        Entity entity = sender.getCommandSenderEntity();
+        HomeController.moveBackHome(entity);
     }
 
     @Override
