+++ src/main/java/io/github/vladocc/getbacktohome/commands/SetHomeCommand.java
@@ -1,8 +1,6 @@
-package com.example.examplemod.commands;
+package io.github.vladocc.getbacktohome.commands;
 
-import com.example.examplemod.HomeController;
-import com.example.examplemod.HomeInfo;
-import com.example.examplemod.HomeProvider;
+import io.github.vladocc.getbacktohome.HomeController;
 import net.minecraft.command.CommandException;
 import net.minecraft.command.ICommand;
 import net.minecraft.command.ICommandSender;
@@ -11,6 +9,8 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.server.MinecraftServer;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.text.TextComponentString;
+import net.minecraft.world.storage.WorldInfo;
+import net.minecraftforge.common.DimensionManager;
 
 import javax.annotation.Nullable;
 import java.util.ArrayList;
@@ -19,22 +19,22 @@ import java.util.List;
 /**
  * Created by Voyager on 06.05.2018.
  */
-public class MoveBackCommand implements ICommand {
+public class SetHomeCommand implements ICommand {
 
     ArrayList<String> allias = new ArrayList<>();
 
-    public MoveBackCommand() {
-        allias.add("back");
+    public SetHomeCommand() {
+        allias.add("sethome");
     }
 
     @Override
     public String getName() {
-        return "home";
+        return "createhome";
     }
 
     @Override
     public String getUsage(ICommandSender sender) {
-        return "home";
+        return "createhome";
     }
 
     @Override
@@ -45,7 +45,7 @@ public class MoveBackCommand implements ICommand {
     @Override
     public void execute(MinecraftServer server, ICommandSender sender, String[] args) throws CommandException {
         Entity entity = sender.getCommandSenderEntity();
-        HomeController.moveBackHome(entity);
+        HomeController.setHome(entity);
     }
 
     @Override
